Fix architecture best-practice violations (DRY, SRP, error handling)
- Extract shared fsys.OSScaffoldFS adapter to eliminate duplication
  between cmd/gc and internal/cityinit test code
- Surface supervisor reload errors via ReloadWarning on InitResult
  and UnregisterResult instead of discarding with _ =
- Move rollback types and functions from service.go into rollback.go
  for single-responsibility separation
- Reword "legacy" comment references to describe current-state intent
