+++ internal/beads/caching_store_test.go
@@ -145,9 +145,8 @@ func TestCachingStorePrimeDoesNotResurrectConcurrentDelete(t *testing.T) {
 	}
 }
 
-func TestCachingStoreCreateRefreshesSparseBead(t *testing.T) {
-	backing := &sparseCreateStore{Store: beads.NewMemStore()}
-	cs := beads.NewCachingStoreForTest(backing, nil)
+func TestCachingStoreCreatePreservesCallerFields(t *testing.T) {
+	cs := beads.NewCachingStoreForTest(beads.NewMemStore(), nil)
 	if err := cs.Prime(context.Background()); err != nil {
 		t.Fatalf("Prime: %v", err)
 	}
@@ -545,18 +544,6 @@ type staleReadAfterUpdateStore struct {
 	mu        sync.Mutex
 	stale     beads.Bead
 	returnOld bool
-}
-
-type sparseCreateStore struct {
-	beads.Store
-}
-
-func (s *sparseCreateStore) Create(b beads.Bead) (beads.Bead, error) {
-	created, err := s.Store.Create(b)
-	if err != nil {
-		return beads.Bead{}, err
-	}
-	return beads.Bead{ID: created.ID, Title: created.Title}, nil
 }
 
 func (s *staleReadAfterUpdateStore) Update(id string, opts beads.UpdateOpts) error {
